+++ src/components/HeatmapAdvancedConfig.tsx
@@ -54,7 +54,10 @@ export function HeatmapAdvancedConfig() {
   return (
     <Accordion type="single" collapsible>
       <AccordionItem value="advanced-config">
-        <AccordionTrigger>Advanced Configuration</AccordionTrigger>
+        <AccordionTrigger>
+          Advanced Configuration&nbsp;
+          <PopoverHelper text="These settings generally do not need to be adjusted and can be ignored." />
+        </AccordionTrigger>
         <AccordionContent>
           <table className="flex flex-row gap-4">
             <tbody>